Separate the clickhouseFactoryInstance so the db client isn't pulled into tests
