+++ apps/webapp/app/routes/admin.api.v1.runs-replication.create.ts
@@ -2,7 +2,7 @@ import { ActionFunctionArgs, json } from "@remix-run/server-runtime";
 import { requireAdminApiRequest } from "~/services/personalAccessToken.server";
 import { z } from "zod";
 import { env } from "~/env.server";
-import { clickhouseFactory } from "~/services/clickhouse/clickhouseFactory.server";
+import { clickhouseFactory } from "~/services/clickhouse/clickhouseFactoryInstance.server";
 import { RunsReplicationService } from "~/services/runsReplicationService.server";
 import {
   getRunsReplicationGlobal,
